requirements.txt
bs4
requests
pypandoc_binary
pandas

